ci: fix Go version, enable golangci-lint v2, tighten config, fix all issues
- Bump CI Go version from 1.22 to 1.25 (eliminates toolchain download)
- Add golangci/golangci-lint-action@v7 to lint job
- Migrate .golangci.yml to v2 format (version, linters.settings, formatters, etc.)
- Remove gosimple (merged into staticcheck), add gosec v2 rule exclusions
- Fix all pre-existing errcheck, gocritic, staticcheck, ineffassign issues
- Proper error handling for type assertions, dead code removal, style fixes

+++ internal/replication/fetcher.go
@@ -338,7 +338,7 @@ func streamReplicaBatches(r io.Reader, totalLen int32, requestedOffset uint64, a
 		if err != nil {
 			return fmt.Errorf("fetcher: parse batch header: %w", err)
 		}
-		batchSize := int32(header.RecordBatchSize())
+		batchSize := header.RecordBatchSize()
 		if batchSize < log.RecordBatchHeaderSize || batchSize > maxReplicaBatchBytes {
 			return fmt.Errorf("fetcher: invalid replica batch size %d", batchSize)
 		}
